enh: email testing utils
