+++ profiles/templates/test_email_body.html
@@ -1,0 +1,5 @@
+{% extends "base_email.html" %}
+
+{% block content %}
+<p style="font-size: 140%">This is a test e-mail.</p>
+{% endblock content %}
