+++ profiles/templates/test_email_body.txt
@@ -1,0 +1,1 @@
+This is a test e-mail.
