+++ profiles/templates/test_email_subject.txt
@@ -1,0 +1,1 @@
+Just a test e-mail.
