+++ winrepo/settings.py
@@ -1,15 +1,3 @@
-"""
-Django settings for winrepo project.
-
-Generated by 'django-admin startproject' using Django 2.0.2.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/2.0/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/2.0/ref/settings/
-"""
-
 import os
 from decouple import config
 
@@ -148,8 +136,10 @@ EMAIL_HOST = config('EMAIL_HOST')
 EMAIL_PORT = config('EMAIL_PORT', default=25, cast=int)
 EMAIL_HOST_USER = config('EMAIL_HOST_USER')
 EMAIL_HOST_PASSWORD = config('EMAIL_HOST_PASSWORD')
-EMAIL_USE_TLS = config('EMAIL_USE_TLS') == 'True'
-EMAIL_USE_SSL = config('EMAIL_USE_SSL') == 'True'
+EMAIL_USE_TLS = config('EMAIL_USE_TLS').lower() == 'true'
+EMAIL_USE_SSL = config('EMAIL_USE_SSL').lower() == 'true'
+
+print('HERE', EMAIL_HOST)
 
 EMAIL_FROM = 'no-reply@winrepo.org'
 EMAIL_SUBJECT_PREFIX = 'WiNRepo - '
